Update index.php
A better method would be to check if the input contains the pre-populated content. The problem with the previous change was that every time someone clicks on this input the contents will be cleared. Even if the user types something and then comes back to edit the field, the text will be cleared. Another problem was that if a user tabs to the input, the content is not being "clicked" on, and therefore does not clear itself. Using onfocus is a better solution.